# backend/swagger-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sesimiz Ol API",
    "version": "1.0.0",
    "description": "Kadınların hikâyelerini güvenle paylaşabilecekleri anonim platform API'si",
    "contact": {
      "name": "Sesimiz Ol Team",
      "url": "https://github.com/sesimiz-ol"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Story": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "title": {"type": "string"},
          "content": {"type": "string"},
          "tags": {"type": "array", "items": {"type": "string"}},
          "authorId": {"type": "string"},
          "createdAt": {"type": "string", "format": "date-time"},
          "likesCount": {"type": "integer"},
          "commentsCount": {"type": "integer"}
        }
      },
...
    }
  },
  "paths": {
...
    "/stories": {
      "get": {
        "tags": ["Stories"],
        "summary": "Hikâyeleri listele",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}},
          {"name": "tag", "in": "query", "schema": {"type": "string"}},
          {"name": "sort", "in": "query", "schema": {"type": "string", "enum": ["newest", "oldest", "popular"]}}
        ],
        "responses": {
          "200": {
            "description": "Hikâye listesi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "stories": {"type": "array", "items": {"$ref": "#/components/schemas/Story"}},
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {"type": "integer"},
                        "page": {"type": "integer"},
                        "pages": {"type": "integer"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Stories"],
        "summary": "Yeni hikâye oluştur",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "content": {"type": "string"},
                  "tags": {"type": "array", "items": {"type": "string"}}
                },
                "required": ["title", "content"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hikâye oluşturuldu",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "story": {"$ref": "#/components/schemas/Story"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}